+++ plugins/WatchStateLoggerPlugin.ts
@@ -29,15 +29,14 @@ export class WatchStateLoggerPlugin {
                 console.log(messages.compilationComplete);
             }
 
-            let emittedFiles = Object
+            const emittedFiles = Object
                 .keys(compilation.assets)
                 .filter(assetKey => compilation.assets[assetKey].emitted);
 
             const chunkFiles = getChunkFiles(compilation);
-
             process.send && process.send(messages.compilationComplete, error => null);
             // Send emitted files so they can be LiveSynced if need be
-            process.send && process.send({ emittedFiles, chunkFiles }, error => null);
+            process.send && process.send({ emittedFiles, chunkFiles, hash: compilation.hash }, error => null);
         });
     }
 }
